+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 - Updated PyTango to v10.0.3
 - Updated `ska-tango-base` to use latest version limited to minor releases (<2.0.0).
 - Removed forked mark on acceptance tests; limited forked tests to tests that use DeviceTestContext.
+- Implemented ConfigureBand command that ingests a JSON string to configure bands on SPFRx.
 
 Version 9.0.0
 *************